# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: "2.0.0"
  title: CCXT REST
  contact:
    name: "Adroit"
    email: "hello@adroit.ph"
    url: "https://adroit.ph/ccxt-rest-contact-us/"
  license:
    name: "MIT"
    url: "https://github.com/franz-see/ccxt-rest/blob/master/LICENSE.txt"
servers:
  - url: 'http://localhost:3000/'
...
paths:
...
  /exchange/{exchangeName}:
    get:
      description: Get the exchange details given the exchangeName and access token in the header
      tags:
        - Exchange Management API
      x-swagger-router-controller: exchange
      operationId: getExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        "404":
          description: Exchange with that name is NOT supported or Exchange with that id does NOT exist
        "500":
          description: If an unexpected error occurred
    post:
      description: Creates an instance of the exchange
      tags:
        - Exchange Management API
      operationId: createExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      requestBody:
        description: The exchange to create.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/exchangeConfig'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
        "503":
          description: Support for exchange is currently broken
      x-swagger-router-controller: exchange
    delete:
      description: Delete the exchange details referenced by access token in the header
      security:
        - bearerAuth: []
      tags:
        - Exchange Management API
      x-swagger-router-controller: exchange
      operationId: deleteExchange
      parameters:
        - $ref: "#/components/parameters/exchangeNameParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        "404":
          description: Exchange with that name is NOT supported
        "500":
          description: If an unexpected error occurred
...
components:
  schemas:
...
    Exchange:
      type: string
      enum: ["_1btcxe","acx","allcoin","anxpro","anybits","bcex","bequant","bibox","bigone","binance","binanceje","bit2c","bitbank","bitbay","bitfinex","bitfinex2","bitflyer","bitforex","bithumb","bitibu","bitkk","bitlish","bitmarket","bitmex","bitsane","bitso","bitstamp","bitstamp1","bittrex","bitz","bl3p","bleutrade","braziliex","btcalpha","btcbox","btcchina","btcexchange","btcmarkets","btctradeim","btctradeua","btcturk","buda","bxinth","ccex","cex","chbtc","chilebit","cobinhood","coinbase","coinbaseprime","coinbasepro","coincheck","coinegg","coinex","coinexchange","coinfalcon","coinfloor","coingi","coinmarketcap","coinmate","coinnest","coinone","coinspot","cointiger","coolcoin","coss","crex24","crypton","cryptopia","deribit","dsx","ethfinex","exmo","exx","fcoin","fcoinjp","flowbtc","foxbit","fybse","fybsg","gateio","gdax","gemini","getbtc","hadax","hitbtc","hitbtc2","huobipro","huobiru","ice3x","independentreserve","indodax","itbit","jubi","kkex","kraken","kucoin","kucoin2","kuna","lakebtc","lbank","liqui","liquid","livecoin","luno","lykke","mandala","mercado","mixcoins","negociecoins","nova","okcoincny","okcoinusd","okex","paymium","poloniex","quadrigacx","rightbtc","southxchange","stronghold","surbitcoin","theocean","therock","tidebit","tidex","uex","upbit","urdubit","vaultoro","vbtc","virwox","xbtce","yobit","zaif","zb"]      
...
    AccessToken:
      required:
        - token
      properties:
        token:
          type: string
          description: >- 
            The JWT token that needs to be added into the 'Authorization'
            header with the 'Bearer ' prefix. For example, given a token of xyz,
            do a secured request with 'Authorization: Bearer xyz'
    ExchangeResponse:
      required:
        - id
        - countries
        - enableRateLimit
        - name
        - private
      type: object
      properties:
        id:
          type: string
          description: >-
            The id of the exchange. When you created the exchange (via
            POST:/exchanges/{exhangeName}), the 'id' parameter there becomes the
            name here
        name:
          type: string
          description: >-
            The name of the exchange. 
        private:
          type: boolean
          description: Whether this exchange is private (has apiKey) or public (no apiKey)
        enableRateLimit:
          type: boolean
          description: >-
            Whether to enable the built in rate limiter or not. The built in
            rate limiter is an approximation of the actual exchange's limit. To
            have a more accurate rate limiting, set this to false and implement
            the rate limiter on your client
          default: true
        countries:
          type: array
          description: The list of countries where this exchange is a member of
          items:
            type: string
        rateLimit:
          type: integer
          description: >-
            A request rate limit in milliseconds. Specifies the required minimal
            delay between two consequent HTTP requests to the same exchange. If
            enableRateLimit is set to false, this would be ignored.
        twofa:
          type: boolean
          description: Whether to enable two factor authentication or not
          default: false
        has:
          $ref: '#/components/schemas/ExchangeHasCapabilities'
        urls:
          type: object
          properties: {}
          description: Collection of URLs this exchange has
...
    ExchangeCapability:
      type: string
      enum:
        - 'true'
        - 'false'
        - emulated
    ExchangeHasCapabilities:
      required:
        - CORS
        - cancelOrder
        - cancelOrders
        - createDepositAddress
        - createLimitOrder
        - createMarketOrder
        - createOrder
        - editOrder
        - fetchBalance
        - fetchBidsAsks
        - fetchClosedOrders
        - fetchCurrencies
        - fetchDepositAddress
        - fetchFundingFees
        - fetchL2OrderBook
        - fetchMarkets
        - fetchMyTrades
        - fetchOHLCV
        - fetchOpenOrders
        - fetchOrder
        - fetchOrderBook
        - fetchOrderBooks
        - fetchOrders
        - fetchTicker
        - fetchTickers
        - fetchTrades
        - fetchTradingFees
        - fetchTradingLimits
        - privateAPI
        - publicAPI
        - withdraw
      type: object
      properties:
        CORS:
          $ref : "#/components/schemas/ExchangeCapability"
        publicAPI:
          $ref : "#/components/schemas/ExchangeCapability"
        privateAPI:
          $ref : "#/components/schemas/ExchangeCapability"
        cancelOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        cancelOrders:
          $ref : "#/components/schemas/ExchangeCapability"
        createDepositAddress:
          $ref : "#/components/schemas/ExchangeCapability"
        createOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        createMarketOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        createLimitOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        editOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchBalance:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchBidsAsks:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchClosedOrders:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchCurrencies:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchDepositAddress:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchFundingFees:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchL2OrderBook:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchMarkets:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchMyTrades:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOHLCV:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOpenOrders:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOrder:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOrderBook:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOrderBooks:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchOrders:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchTicker:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchTickers:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchTrades:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchTradingFees:
          $ref : "#/components/schemas/ExchangeCapability"
        fetchTradingLimits:
          $ref : "#/components/schemas/ExchangeCapability"
        withdraw:
          $ref : "#/components/schemas/ExchangeCapability"
    exchangeConfig:
      required:
        - id
      type: object
      properties:
        id:
          type: string
          description: >-
            The unique identifier for this exchange. This would be used as the
            'exchangeId' of the other API calls
        apiKey:
          type: string
          description: >-
            The API key you got from the exchange itself. This with the secret
            is what will allow you to access the exchange
        secret:
          type: string
          description: >-
            The Secret key you got from the exchange itself. This with the
            apiKey is what will allow you to access the exchange
        enableRateLimit:
          type: boolean
          description: >-
            Whether to enable the built in rate limiter or not. The built in
            rate limiter is an approximation of the actual exchange's limit. To
            have a more accurate rate limiting, set this to false and implement
            the rate limiter on your client
          default: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  
  parameters:
...
    exchangeNameParam:
      name: exchangeName
      in: path
      description: >-
        The name of the exchange. Possible values are any of the result of
        GET:/exchanges.
      required: true
      style: simple
      explode: false
      schema:
        $ref : "#/components/schemas/Exchange"